# openapi.yaml
...
openapi: 3.1.0
info:
  title: Lentera — LMS Wakaf Salman API
  version: 1.0.0
  description: |-
    Kontrak berdasarkan implementasi Flask di repository. Gunakan Bearer JWT dari POST /api/auth/login. Logout tercatat melalui POST /api/auth/logout; token baru yang terkait sesi login tersebut ditolak setelah logout. Role dan status akun dibaca ulang dari database pada setiap request.

    Path learning memakai /api/course (singular); assessment memakai /api/courses (plural). ID berasal dari database, bukan ID contoh/frontend. Daftar katalog/user/enrollment tidak memiliki pagination; hanya activity mendukung pagination.

    Payload JSON memakai Content-Type: application/json. Field request yang tidak dikenal ditolak kecuali login/register yang mengabaikan field tambahan. Input teks umumnya di-trim. Global request limit 3 MiB; endpoint profil memakai batas 2 MiB + 64 KiB. Error aplikasi umumnya {message}; error framework/proxy (termasuk 413 di luar auth) dapat berupa HTML. Timestamp identitas/learning dapat tanpa timezone, sedangkan timestamp assessment/discussion memakai UTC dengan Z.

    Endpoint Development bersifat opsional dan selalu tidak tersedia pada production. Spesifikasi ini tidak menambahkan route HTTP baru atau Swagger UI.
servers:
- url: http://localhost:5000
  description: Flask development lokal
- url: http://localhost:8000
  description: Docker API lokal; sesuaikan API_PORT
- url: https://{hostname}
...
  variables:
    hostname:
      default: api.example.com
...
security:
- bearerAuth: []
paths:
...
  /api/course:
    get:
      tags:
      - Course
      summary: Daftar course
      operationId: get_api_course
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  course:
                    type: array
                    items:
                      $ref: '#/components/schemas/Course'
                required:
                - course
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
      description: User hanya melihat data aktif dengan kategori aktif. Admin melihat semua status. Tanpa pagination;
        urut ID menurun.
    post:
      tags:
      - Course
      summary: Tambah course
      operationId: post_api_course
      responses:
        '201':
          description: Berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  course:
                    $ref: '#/components/schemas/Course'
                required:
                - message
                - course
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
      x-required-role: admin
      description: 'Hanya admin aktif. '
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseCreate'
...
  /api/activity:
    get:
      tags:
      - Reporting
      summary: Log aktivitas
      operationId: get_api_activity
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  activities:
                    type: array
                    items:
                      $ref: '#/components/schemas/Activity'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                required:
                - activities
                - pagination
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
      x-required-role: admin
      description: Hanya admin aktif. Urut waktu_dimulai lalu id_activity menurun. Login sukses sudah mencatat activity;
        endpoint ini hanya membaca event tersimpan.
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 1000000
          default: 1
      - name: per_page
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
      - name: id_user
        in: query
        schema:
          type: integer
          format: int64
          minimum: 1
          maximum: 4294967295
      - name: jenis_aktivitas
        in: query
        schema:
          type: string
          enum:
          - login
          - buka_course
          - buka_materi
          - selesai_materi
          - selesai_course
          - mulai_test
          - selesai_test
          - logout
          - sesi
          - buka_dashboard
          - putar_video
          - kirim_komentar
      - {name: search, in: query, description: Nama/email user atau judul course, schema: {type: string, maxLength: 200}}
      - {name: user, in: query, description: Nama/email user, schema: {type: string, maxLength: 200}}
      - {name: course, in: query, description: Judul course, schema: {type: string, maxLength: 200}}
      - {name: date_from, in: query, description: Awal hari inklusif dalam WIB, schema: {type: string, format: date}}
      - {name: date_to, in: query, description: Akhir hari inklusif dalam WIB, schema: {type: string, format: date}}
  /api/dashboard:
    get:
      tags:
      - Reporting
      summary: Statistik identitas
      operationId: get_api_dashboard
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
      x-required-role: admin
      description: 'Hanya admin aktif. '
  /api/leaderboard:
    get:
      tags:
      - Reporting
      summary: Status ketersediaan leaderboard
      operationId: get_api_leaderboard
      responses:
        '501':
          description: Berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaderboardUnavailable'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
      description: Belum tersedia. Tidak mengembalikan ranking atau array kosong seolah-olah ranking sudah diimplementasikan.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT HS256 dari login; Authorization: Bearer <token>. Berlaku 8 jam.'
  responses:
    Error400:
      description: Payload atau parameter tidak valid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error401:
      description: Token hilang/invalid/kedaluwarsa, akun tidak ditemukan, atau kredensial login salah.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error403:
      description: Akun nonaktif, role tidak sesuai, bukan pemilik, atau prasyarat akses belum terpenuhi.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error404:
      description: Data tidak ditemukan atau tidak dapat dilihat oleh pengguna.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error409:
      description: Konflik data, relasi, konfigurasi test, atau status operasi.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
          example: Data tidak valid
      required:
      - message
...
    CourseCreate:
      type: object
      properties:
        judul_course:
          type: string
          minLength: 1
          maxLength: 150
        deskripsi:
          anyOf:
          - type: string
            description: Maksimal 65535 byte UTF-8.
          - type: 'null'
        id_kategori:
          type: integer
          format: int64
          minimum: 1
          maximum: 4294967295
        key_point:
          anyOf:
          - type: string
            description: Maksimal 65535 byte UTF-8.
          - type: 'null'
        credit:
          anyOf:
          - type: string
            maxLength: 150
          - type: 'null'
        thumbnail:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
        status:
          type: string
          enum:
          - aktif
          - nonaktif
      required:
      - judul_course
      - id_kategori
      additionalProperties: false
      description: Status default aktif.
...
    Course:
      type: object
      properties:
        id_course:
          type: integer
          format: int64
          minimum: 1
        judul_course:
          type: string
          minLength: 1
          maxLength: 150
        deskripsi:
          anyOf:
          - type: string
            description: Maksimal 65535 byte UTF-8.
          - type: 'null'
        id_kategori:
          type: integer
          format: int64
          minimum: 1
          maximum: 4294967295
        key_point:
          anyOf:
          - type: string
            description: Maksimal 65535 byte UTF-8.
          - type: 'null'
        credit:
          anyOf:
          - type: string
            maxLength: 150
          - type: 'null'
        thumbnail:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
        status:
          type: string
          enum:
          - aktif
          - nonaktif
        tanggal_dibuat:
          anyOf:
          - type: string
            description: Tanggal/waktu ISO 8601 tanpa timezone, mengikuti waktu server.
            example: '2026-10-06T10:30:00'
          - type: 'null'
        tanggal_diperbarui:
          anyOf:
          - type: string
            description: Tanggal/waktu ISO 8601 tanpa timezone, mengikuti waktu server.
            example: '2026-10-06T10:30:00'
          - type: 'null'
      required:
      - id_course
      - judul_course
      - deskripsi
      - id_kategori
      - key_point
      - credit
      - thumbnail
      - status
      - tanggal_dibuat
      - tanggal_diperbarui
...
    Activity:
      type: object
      properties:
        id_activity:
          type: integer
          format: int64
          minimum: 1
        id_user:
          type: integer
          format: int64
          minimum: 1
        id_course:
          anyOf:
          - type: integer
            format: int64
            minimum: 1
          - type: 'null'
        id_materi:
          anyOf:
          - type: integer
            format: int64
            minimum: 1
          - type: 'null'
        id_penilaian:
          anyOf:
          - type: integer
            format: int64
            minimum: 1
          - type: 'null'
        jenis_aktivitas:
          type: string
          enum:
          - login
          - buka_course
          - buka_materi
          - selesai_materi
          - selesai_course
          - mulai_test
          - selesai_test
          - logout
          - sesi
          - buka_dashboard
          - putar_video
          - kirim_komentar
        user_name: {type: string}
        course_title: {type: [string, 'null']}
        material_title: {type: [string, 'null']}
        time_basis: {type: [string, 'null'], description: UTC untuk event baru; null untuk legacy}
        duration_status: {type: string, enum: [completed, confirmed_partial, recorded, unavailable]}
        durasi:
          anyOf:
          - type: integer
            minimum: 0
            maximum: 4294967295
          - type: 'null'
        waktu_dimulai:
          anyOf:
          - type: string
            description: Event baru UTC dengan Z; legacy tanpa zona dipertahankan.
            example: '2026-10-06T10:30:00'
          - type: 'null'
        waktu_selesai:
          anyOf:
          - type: string
            description: Batas waktu terakhir terkonfirmasi; untuk tracking selesai adalah waktu end. UTC dengan Z untuk event baru.
            example: '2026-10-06T10:30:00'
          - type: 'null'
      required:
      - id_activity
      - id_user
      - id_course
      - id_materi
      - id_penilaian
      - jenis_aktivitas
      - durasi
      - waktu_dimulai
      - waktu_selesai
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        per_page:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        pages:
          type: integer
          minimum: 0
      required:
      - page
      - per_page
      - total
      - pages
    Dashboard:
      type: object
      properties:
        identity:
          type: object
          properties:
            total_users:
              type: integer
              minimum: 0
            active_users:
              type: integer
              minimum: 0
            inactive_users:
              type: integer
              minimum: 0
            admins:
              type: integer
              minimum: 0
            regular_users:
              type: integer
              minimum: 0
          required:
          - total_users
          - active_users
          - inactive_users
          - admins
          - regular_users
        learning:
          type: 'null'
        assessment:
          type: 'null'
        dependencies:
          type: array
          items:
            type: string
      required:
      - identity
      - learning
      - assessment
      - dependencies
      description: Statistik learning dan assessment belum diimplementasikan, sehingga selalu null.
    LeaderboardUnavailable:
      type: object
      properties:
        status:
          type: string
          enum:
          - pending_dependencies
        message:
          type: string
        dependencies:
          type: array
          items:
            type: string
      required:
      - status
      - message
      - dependencies
...